+++ challenge7/service/book.go
@@ -6,7 +6,7 @@ type BookService interface {
 	CreateBook(in model.Book) (res model.Book, err error)
 	GetAllBook() (res []model.Book, err error)
 	GetBookByID(id int64) (res model.Book, err error)
-	UpdateBook(id int, data model.Book) error
+	UpdateBook(id int64, data model.Book) error
 	DeleteBook(id int64) (err error)
 }
 
@@ -29,7 +29,7 @@ func (s *Service) GetBookByID(id int64) (res model.Book, err error) {
 	return s.repo.GetBookByID(id)
 }
 
-func (s *Service) UpdateBook(id int, data model.Book) error {
+func (s *Service) UpdateBook(id int64, data model.Book) error {
 	return s.repo.UpdateBook(id, data)
 }
